+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Add eslint
 
+### Fixed
+- Switch to GraphQL query variables to prevent string escaping issues
+
 ## [0.4.0] - 2023-05-12
 ### Changed
 - Rename organization